requirements.txt
Django==4.1.2
Pillow
flake8
djangorestframework
markdown
autopep8